+++ src/utils/index.ts
@@ -16,4 +16,4 @@
  */
 
 export * from "./packageUtils";
-export * from "./slslDependencies";
+export * from "./slsDependencies";
